splitting redux slices
