+++ redux/src/store/auth.js
@@ -1,0 +1,17 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+const authSlice = createSlice({
+	name: 'auth',
+	initialState: { isAuthenticated: false },
+	reducers: {
+		login(state) {
+			state.isAuthenticated = true;
+		},
+		logout(state) {
+			state.isAuthenticated = false;
+		},
+	},
+});
+
+export const authActions = authSlice.actions;
+export default authSlice.reducer;
