+++ redux/src/store/index.js
@@ -1,46 +1,15 @@
-import { createSlice, configureStore } from '@reduxjs/toolkit';
+import { configureStore } from '@reduxjs/toolkit';
+import counterSlice from './counter';
+import authSlice from './auth';
 
-const initialCounterState = { counter: 0, showCounter: true };
-
-const counterSlice = createSlice({
-	name: 'counter',
-	initialState: initialCounterState,
-	reducers: {
-		increment(state) {
-			state.counter++;
-		},
-		decrement(state) {
-			state.counter--;
-		},
-		increase(state, action) {
-			state.counter = state.counter + action.payload;
-		},
-		toggleCounter(state) {
-			state.showCounter = !state.showCounter;
-		},
-	},
-});
-
-const authSlice = createSlice({
-  name: 'auth',
-  initialState: { isAuthenticated: false },
-  reducers: {
-    login(state) {
-      state.isAuthenticated = true;
-    },
-    logout(state) {
-      state.isAuthenticated = false;
-    },
-  },
-});
 
 const store = configureStore({
 	reducer: {
-    counter: counterSlice.reducer,
-    auth: authSlice.reducer
+    counter: counterSlice,
+    auth: authSlice,
   },
 });
 
-export const counterActions = counterSlice.actions;
-export const authActions = authSlice.actions;
+
+
 export default store;
